feat: handle author deletion

+++ src/authors/authors.controller.ts
@@ -2,10 +2,15 @@ import { Controller, Get, Post, Body, Patch, Param, Delete, NotFoundException, P
 import { AuthorsService } from './authors.service';
 import { CreateAuthorDto } from './dto/create-author.dto';
 import { UpdateAuthorDto } from './dto/update-author.dto';
+import { BooksService } from 'src/books/books.service';
+import { AuthorHasExistingBooksException } from './authors.exception';
 
 @Controller('authors')
 export class AuthorsController {
-  constructor(private readonly authorsService: AuthorsService) {}
+  constructor(
+    private readonly authorsService: AuthorsService,
+    private readonly booksService: BooksService
+  ) {}
 
   @Post()
   create(@Body(new ValidationPipe()) createAuthorDto: CreateAuthorDto) {
@@ -37,10 +42,12 @@ export class AuthorsController {
 
   @Delete(':id')
   remove(@Param('id', ParseIntPipe) id: number) {
-    try {
-      return this.authorsService.remove(id);
-    } catch (err) {
-      throw new NotFoundException();
+    const booksByAuthor = this.booksService.findAuthorsBooks(id);
+
+    if (booksByAuthor.length > 0) {
+      throw new AuthorHasExistingBooksException();
     }
+
+    return this.authorsService.remove(id);
   }
 }
